Update for domain names...

+++ run_client.sh
@@ -1,7 +1,14 @@
+#!/bin/bash
+
+if [ "$#" -ne 3 ]; then
+	echo "Invalid number of parameters, $0 ServerIP DomainName ProtocolType"
+	exit 1
+fi
+
 rm -rf DNS_UDP_Client
 g++ -std=c++11 DNS_UDP_Client.cpp -o DNS_UDP_Client -lpthread
 
 rm -rf file.pdf
 
-# DNS_UDP_Client serverHostname fileName protocolType
-./DNS_UDP_Client 10.0.2.5 file.pdf 2
+# DNS_UDP_Client ServerIP DomainName ProtocolType
+./DNS_UDP_Client $1 $2 $3
